+++ di/extractor.go
@@ -1,6 +1,0 @@
-package di
-
-// Extractor
-type Extractor interface {
-	Extract(params ExtractParams) error
-}