+++ di/internal/graphkv/directed_graph.go
@@ -1,34 +1,34 @@
-package dag
+package graphkv
 
-// DirectedGraph is a graph supporting directed edges between nodes.
-type DirectedGraph struct {
+// directedGraph is a graph supporting directed edges between nodes.
+type directedGraph struct {
 	*graph
 	edges *directedEdgeList
 }
 
-// NewDirectedGraph creates a graph of nodes with directed edges.
-func NewDirectedGraph() *DirectedGraph {
-	return &DirectedGraph{
+// newDirectedGraph creates a graph of nodes with directed edges.
+func newDirectedGraph() *directedGraph {
+	return &directedGraph{
 		graph: newGraph(),
 		edges: newDirectedEdgeList(),
 	}
 }
 
 // Copy returns a clone of the directed graph.
-func (g *DirectedGraph) Copy() *DirectedGraph {
-	return &DirectedGraph{
+func (g *directedGraph) Copy() *directedGraph {
+	return &directedGraph{
 		graph: g.graph.Copy(),
 		edges: g.edges.Copy(),
 	}
 }
 
 // EdgeCount returns the number of direced edges between nodes.
-func (g *DirectedGraph) EdgeCount() int {
+func (g *directedGraph) EdgeCount() int {
 	return g.edges.Count()
 }
 
 // AddEdge adds the edge to the graph.
-func (g *DirectedGraph) AddEdge(from Node, to Node) {
+func (g *directedGraph) AddEdge(from Key, to Key) {
 	// prevent adding an edge referring to missing nodes
 	if !g.NodeExists(from) {
 		g.AddNode(from)
@@ -41,12 +41,12 @@ func (g *DirectedGraph) AddEdge(from Node, to Node) {
 }
 
 // RemoveEdge removes the edge from the graph.
-func (g *DirectedGraph) RemoveEdge(from Node, to Node) {
+func (g *directedGraph) RemoveEdge(from Key, to Key) {
 	g.edges.Remove(from, to)
 }
 
 // HasEdges determines whether the graph contains any edges to or from the node.
-func (g *DirectedGraph) HasEdges(node Node) bool {
+func (g *directedGraph) HasEdges(node Key) bool {
 	if g.HasIncomingEdges(node) {
 		return true
 	}
@@ -54,50 +54,50 @@ func (g *DirectedGraph) HasEdges(node Node) bool {
 }
 
 // EdgeExists checks whether the edge exists within the graph.
-func (g *DirectedGraph) EdgeExists(from Node, to Node) bool {
+func (g *directedGraph) EdgeExists(from Key, to Key) bool {
 	return g.edges.Exists(from, to)
 }
 
 // HasIncomingEdges checks whether the graph contains any directed
 // edges pointing to the node.
-func (g *DirectedGraph) HasIncomingEdges(node Node) bool {
+func (g *directedGraph) HasIncomingEdges(node Key) bool {
 	return g.edges.HasIncomingEdges(node)
 }
 
 // IncomingEdges returns the nodes belonging to directed edges pointing
 // towards the specified node.
-func (g *DirectedGraph) IncomingEdges(node Node) []Node {
+func (g *directedGraph) IncomingEdges(node Key) []Key {
 	return g.edges.IncomingEdges(node)
 }
 
 // IncomingEdgeCount returns the number of edges pointing from the specified
 // node (indegree).
-func (g *DirectedGraph) IncomingEdgeCount(node Node) int {
+func (g *directedGraph) IncomingEdgeCount(node Key) int {
 	return g.edges.IncomingEdgeCount(node)
 }
 
 // HasOutgoingEdges checks whether the graph contains any directed
 // edges pointing from the node.
-func (g *DirectedGraph) HasOutgoingEdges(node Node) bool {
+func (g *directedGraph) HasOutgoingEdges(node Key) bool {
 	return g.edges.HasOutgoingEdges(node)
 }
 
 // OutgoingEdges returns the nodes belonging to directed edges pointing
 // from the specified node.
-func (g *DirectedGraph) OutgoingEdges(node Node) []Node {
+func (g *directedGraph) OutgoingEdges(node Key) []Key {
 	return g.edges.OutgoingEdges(node)
 }
 
 // OutgoingEdgeCount returns the number of edges pointing from the specified
 // node (outdegree).
-func (g *DirectedGraph) OutgoingEdgeCount(node Node) int {
+func (g *directedGraph) OutgoingEdgeCount(node Key) int {
 	return g.edges.OutgoingEdgeCount(node)
 }
 
 // RootNodes finds the entry-point nodes to the graph, i.e. those without
 // incoming edges.
-func (g *DirectedGraph) RootNodes() []Node {
-	results := make([]Node, 0)
+func (g *directedGraph) RootNodes() []Key {
+	results := make([]Key, 0)
 	for _, node := range g.Nodes() {
 		if !g.HasIncomingEdges(node) {
 			results = append(results, node)
@@ -107,8 +107,8 @@ func (g *DirectedGraph) RootNodes() []Node {
 }
 
 // IsolatedNodes finds independent nodes in the graph, i.e. those without edges.
-func (g *DirectedGraph) IsolatedNodes() []Node {
-	results := make([]Node, 0)
+func (g *directedGraph) IsolatedNodes() []Key {
+	results := make([]Key, 0)
 	for _, node := range g.Nodes() {
 		if !g.HasEdges(node) {
 			results = append(results, node)
@@ -119,10 +119,10 @@ func (g *DirectedGraph) IsolatedNodes() []Node {
 
 // AdjacencyMatrix returns a matrix indicating whether pairs of nodes are
 // adjacent or not within the graph.
-func (g *DirectedGraph) AdjacencyMatrix() map[Node]map[Node]bool {
-	matrix := make(map[Node]map[Node]bool, g.NodeCount())
+func (g *directedGraph) AdjacencyMatrix() map[Key]map[Key]bool {
+	matrix := make(map[Key]map[Key]bool, g.NodeCount())
 	for _, a := range g.Nodes() {
-		matrix[a] = make(map[Node]bool, g.NodeCount())
+		matrix[a] = make(map[Key]bool, g.NodeCount())
 
 		for _, b := range g.Nodes() {
 			matrix[a][b] = g.EdgeExists(a, b)
@@ -133,7 +133,7 @@ func (g *DirectedGraph) AdjacencyMatrix() map[Node]map[Node]bool {
 
 // RemoveTransitives removes any transitive edges so that as fewest possible
 // edges exist while matching the reachability of the original graph.
-func (g *DirectedGraph) RemoveTransitives() {
+func (g *directedGraph) RemoveTransitives() {
 	for _, a := range g.Nodes() {
 		for _, b := range g.Nodes() {
 			if !g.EdgeExists(a, b) {